requirement.txt
requests
pytz
python-quoine
pysher
pubnub
git+https://github.com/bitbankinc/python-bitbankcc.git